+++ examples/02-routing/spec.js
@@ -4,13 +4,23 @@ describe('React server side rendering', function () {
         expect(element.all(by.css('ul li')).count()).toEqual(5);
     });
 
-    it('title should be main', function () {
+    it('body title should be main', function () {
         browser.get('main');
         expect(element(by.css('div h1')).getInnerHtml()).toMatch(/Main/);
+    });
+
+    it('html title should be Main Page', function () {
+        expect(element(by.css('title')).getInnerHtml()).toBe('Main Page');
     });
 });
 
 describe('React client side binding', function () {
+    it('should handle document.title correctly', function () {
+        browser.driver.executeScript('return document.title').then(function (value) {
+            expect(value).toBe('Main Page');
+        });
+    });
+
     it('should handle button click', function () {
         browser.get('main');
         browser.driver.executeScript('return window.test=1');
@@ -21,8 +31,42 @@ describe('React client side binding', function () {
             });
         });
         expect(element(by.css('div span')).getInnerHtml()).toMatch(/12300/);
+    });
+
+    it('should block page loading and use history.pushState', function () {
         browser.driver.executeScript('return window.test').then(function (value) {
             expect(value).toBe(1);
+        });
+    });
+
+    it('should change title correctly after page changed', function () {
+        browser.driver.executeScript('return document.title').then(function (value) {
+            expect(value).toMatch(/123/);
+        });
+    });
+
+    it('should handle browser back button', function () {
+        // prevent safari issue by javascript history.go(-1)
+        // http://stackoverflow.com/questions/27240969/unknown-error-on-safari-history-navigation-using-protractor-webdrivers
+        browser.driver.executeScript('return (window.test=2) && (history.go(-1))');
+
+        browser.wait(function () {
+            return browser.driver.getCurrentUrl().then(function (url) {
+                return url.match(/main/);
+            });
+        });
+        expect(element(by.css('div h1')).getInnerHtml()).toMatch(/Main/);
+    });
+
+    it('should not cause page changed when back', function () {
+        browser.driver.executeScript('return window.test').then(function (value) {
+            expect(value).toBe(2);
+        });
+    });
+
+    it('should change title correctly after page back', function () {
+        browser.driver.executeScript('return document.title').then(function (value) {
+            expect(value).toBe('Main Page');
         });
     });
 });