+++ examples/shared_testconf.js
@@ -16,7 +16,7 @@ var config = {
     {browserName: 'internet explorer', version: 11, platform: 'Windows 8.1'},
     {browserName: 'internet explorer', version: 10, platform: 'Windows 8'},
     {browserName: 'internet explorer', version: 9, platform: 'Windows 7'},
-    {browserName: 'internet explorer', version: 8, platform: 'Windows XP'},
+    {browserName: 'internet explorer', version: 8, platform: 'Windows 7'},
     {browserName: 'Android', version:'4.4', platform: 'Linux'},
     {browserName: 'Android', version:'4.1', platform: 'Linux'},
     {browserName: 'Android', version:'4.0', platform: 'Linux'},