+++ extra/FluxexHtml.jsx
@@ -2,6 +2,7 @@
 
 var React = require('react'),
     Fluxex = require('fluxex'),
+    InitScript = Fluxex.InitScript,
 
 FluxexHtml = React.createClass({
     mixins: [
@@ -20,9 +21,7 @@ FluxexHtml = React.createClass({
          </head>
          <body onClick={this.handleClickLink}>
           {this.props.children}
-          <script src="http://cdn.jsdelivr.net/polyfills/polyfill.js"></script>
-          <script src="/static/js/main.js"></script>
-          <script dangerouslySetInnerHTML={{__html: this.getInitScript()}}></script>
+          <InitScript />
          </body>
         </html> 
         );