+++ extra/dummy.js
@@ -1,0 +1,3 @@
+// This is a dummy file to remove client unused codes
+// check extra/gulpfile.js to see alias settings
+// Search for aliasify