+++ extra/gulpfile.js
@@ -40,8 +40,11 @@ configs = {
     aliasify: {
         aliases: {
             request: 'browser-request',
+            'fluxex/extra/history': 'html5-history-api',
             'fluxex/extra/polyfill': 'babelify/polyfill',
-            'core-js/shim': 'core-js/client/shim'
+            'fluxex/extra/polyfill-ie8': 'fluxex/extra/polyfill-ie8-client',
+            'core-js/shim': 'core-js/client/shim',
+            'body-parser': 'fluxex/extra/dummy'
         }
     },
     babelify: {