+++ extra/polyfill-ie8.js
@@ -1,0 +1,5 @@
+'use strict';
+// This is a dummy file to enable eventListener polyfill for IE8
+// Do this in your first line of fluxex.js to enable polyfill:
+// require('fluxex/extra/polyfill-ie8');
+// our gulp task will alias it to 'fluxex/extra/polyfill-ie8-client'